backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.25.0
supabase>=2.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
aiofiles>=23.0.0

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0

# Google Gemini Vision AI
google-generativeai>=0.4.0

# MQTT for IoT sensors (optional - for production)
aiomqtt>=2.0.0

# Groq Cloud LLaMA
groq>=0.5.0

# Firebase Realtime Database
firebase-admin>=6.0.0
